+++ Protocol/Line.cs
@@ -13,16 +13,16 @@ namespace SysCAD.Protocol
   public class Line : Element
   {
 
-    public Single x1;
-    public Single x2;
-    public Single y1;
-    public Single y2;
+    public Double x1;
+    public Double x2;
+    public Double y1;
+    public Double y2;
 
     public Line()
     {
     }
 
-    public Line(Single x1, Single y1, Single x2, Single y2)
+    public Line(Double x1, Double y1, Double x2, Double y2)
     {
       this.x1 = x1;
       this.y1 = y1;
